Remove overview from NFT page
