+++ src/Pages/OpenSea/index.tsx
@@ -1,45 +1,44 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { Row, Col } from "react-bootstrap";
+import { Row, Col, Button } from "react-bootstrap";
 import { AppState } from "../../state";
-import ValueCard from "../../components/ValueCard";
-import OpenSeaContainer from "../../components/OpenSeaContainer";
+import CustomCard, { CustomHeader, CustomTitle } from "../../components/CustomCard";
+import OpenSeaCard from "../../components/OpenSeaCard";
 
 const OpenSea = () => {
-  const overview = useSelector((state: AppState) => state.balances.overview);
+  const assets = useSelector((state: AppState) => state.opensea.assets);
+  const assetList = assets.map((asset, index) => (
+    <div className='col mb-8' key={index}>
+      <OpenSeaCard asset={asset} />
+    </div>
+  ));
 
   return (
     <>
       <Row>
-        <Col xs={12} md={4}>
-          <ValueCard
-            className={"gutter-b"}
-            title={"Total Assets"}
-            value={overview.deposits.total + overview.wallet.total}
-          />
-        </Col>
-        <Col xs={12} md={4}>
-          <ValueCard
-            className={"gutter-b"}
-            title={"Total Debt"}
-            value={overview.debts.total}
-          />
-        </Col>
-        <Col xs={12} md={4}>
-          <ValueCard
-            className={"gutter-b"}
-            title={"Net Worth"}
-            value={
-              overview.deposits.total +
-              overview.wallet.total -
-              overview.debts.total
-            }
-          />
-        </Col>
-      </Row>
-      <Row>
         <Col xs={12}>
-          <OpenSeaContainer />
+          <CustomCard className="test">
+            <CustomHeader className="card-header d-flex align-items-center justify-content-between">
+                <CustomTitle className="card-title">
+                  Non-Fungible Tentacles
+                </CustomTitle>
+                <Button
+                  type='button'
+                  className='btn btn-success font-weight-bolder font-size-lg px-9 py-3'
+                  href='https://opensea.io/accounts/octofi'>
+                    Visit OpenSea
+                </Button>
+            </CustomHeader>
+            <div className="card-body">
+              {!!assets.length ? (
+                <div className='row row-cols-1 row-cols-sm-2 row-cols-md-3 row-cols-lg-4'>
+                  {assetList}
+                </div>
+              ) : (
+                <p className='h3'>No items available at this time.</p>
+              )}
+            </div>
+          </CustomCard>
         </Col>
       </Row>
     </>
